+++ flocker/volume/testtools.py
@@ -14,6 +14,7 @@ from characteristic import attributes
 from twisted.python.filepath import FilePath
 from twisted.internet.task import Clock
 from twisted.internet import reactor
+from twisted.trial.unittest import SynchronousTestCase
 
 from ..common import ProcessNode
 from ._ipc import RemoteVolumeManager
@@ -151,3 +152,66 @@ def create_realistic_servicepair(test):
                                  to_config)
     return ServicePair(from_service=from_service, to_service=to_service,
                        remote=remote)
+
+
+def make_volume_options_tests(make_options, extra_arguments=None):
+    """
+    Make a ``TestCase`` to test the ``VolumeService`` specific arguments added
+    to an ``Options`` class by the ``flocker_volume_options`` class decorator.
+
+    :param make_options: A zero-argument callable which will be called to
+        produce the ``Options`` instance under test.
+    :param extra_arguments: An optional ``list`` of non-VolumeService related
+        arguments which are required by the ``Options`` instance under test.
+    :return: A ``SynchronousTestCase``.
+    """
+    if extra_arguments is None:
+        extra_arguments = []
+
+    def parseOptions(options, argv):
+        options.parseOptions(argv + extra_arguments)
+
+    class VolumeOptionsTests(SynchronousTestCase):
+        """
+        Tests for ``Options`` subclasses decorated with
+        ``flocker_volume_options``.
+        """
+        def test_default_config(self):
+            """
+            By default the config file is ``b'/etc/flocker/volume.json'``.
+            """
+            options = make_options()
+            parseOptions(options, [])
+            self.assertEqual(
+                FilePath(b"/etc/flocker/volume.json"), options["config"])
+
+        def test_config(self):
+            """
+            The options class accepts a ``--config`` parameter.
+            """
+            path = b"/foo/bar"
+            options = make_options()
+            parseOptions(options, [b"--config", path])
+            self.assertEqual(FilePath(path), options["config"])
+
+        def test_pool(self):
+            """
+            The options class accepts a ``--pool`` parameter.
+            """
+            pool = b"foo-bar"
+            options = make_options()
+            parseOptions(options, [b"--pool", pool])
+            self.assertEqual(pool, options["pool"])
+
+        def test_mountpoint(self):
+            """
+            The options class accepts a ``--mountpoint`` parameter.
+            """
+            mountpoint = b"/bar/baz"
+            options = make_options()
+            parseOptions(options, [b"--mountpoint", mountpoint])
+            self.assertEqual(mountpoint, options["mountpoint"])
+
+    dummy_options = make_options()
+    VolumeOptionsTests.__name__ = dummy_options.__class__.__name__ + "Tests"
+    return VolumeOptionsTests