+++ docs/advanced/whatsnew.rst
@@ -1,3 +1,11 @@
+==============================
+What's New in the Next Release
+==============================
+
+* Added support for environment variables in application configuration.
+
+  * See :doc:`./configuration` for more information.
+
 ==================
 What's New in v0.1
 ==================